pyproject.toml
[project]
name = "mol-mcp"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "fast-agent-mcp",
    "fastmcp>=2.12.4",
    "rdkit>=2025.9.1",
    "ase>=3.26.0",
    "tblite>=0.5.0 ; platform_system == 'Darwin'",
    "tblite==0.4.0 ; platform_system != 'Darwin'",
    "molify>=0.2.2",
    "typer>=0.19.2",
    "python-dotenv>=1.1.1",
]

[project.scripts]
mol-mcp = "molmcp.__main__:main"
smls-mcp = "molmcp.tools.smileymcp.server:main"
geom-mcp = "molmcp.tools.qcmcp.server:main"
pack-mcp = "molmcp.tools.packmcp.server:main"
calc-mcp = "molmcp.tools.calcmcp.server:main"


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"molmcp" = ["src/molmcp/agents/*/*.yaml"]

[tool.setuptools.packages.find]
namespaces = true
where = ["src"]

[tool.uv]
index-strategy="unsafe-first-match"
